fea: Remove the cast and provide a default implementation in the mcp_protocol trait.
